phpbb2/requirements.txt
Django==1.4
South==0.7.5
django-crispy-forms==1.1.4
django-recaptcha==0.0.4
django-registration==0.8
#MySQL-python
